把long long改为long，两者无区别

+++ ebox_stm32f103RCT6_VS/ebox_stm32f103RCT6_VS/common/edriver/encoder_exti.cpp
@@ -50,7 +50,7 @@ void EncoderExti::begin()
 	extiB.interrupt(ENABLE);
 }
 
-long long EncoderExti::getPosition()
+long EncoderExti::getPosition()
 {
 	return position;
 }
